refactor: Relocate chatbot packages and remove exec-process pending
