+++ internal/runtime/chatbot/discord/attachments.go
