+++ internal/runtime/tui/commandTelegram.go
@@ -9,7 +9,7 @@ import (
 
 	tea "github.com/charmbracelet/bubbletea"
 
-	"github.com/pardnchiu/agenvoy/internal/runtime/telegram"
+	"github.com/pardnchiu/agenvoy/internal/runtime/chatbot/telegram"
 	"github.com/pardnchiu/agenvoy/internal/session/config"
 	go_bot_telegram "github.com/pardnchiu/go-bot/telegram"
 	"github.com/pardnchiu/go-pkg/filesystem/keychain"
